Fixed Railtie to only load helper for UI controller
Necessary to prevent the following error when a Rails application is
initialized and uses the `ActionController::API` for a base class:

    NoMethodError:
      undefined method `helper_method' for ActionController::API:Class

Normally, when configuring Rails, you can specify if your application
is an UI or API only (UI is the default) and Rails will auto-filter
initializers which don't support the one or the other. Unfortunately,
the `:action_controller` hook is a superset hook which causes issues
when working in a hybrid application (i.e. UI + API). Here's a better
illustration:

- `:action_controller` - `ApplicationController` and
  `ActionController::API`.
- `:action_controller_base` - `ApplicationController` only.
- `:action_controller_api` - `ActionController::API` only.