Refactored implementation to use anonymous block forwarding
Necessary to reduce typing and redundant code when we can use the
shorthand anonymous block forwarding syntax introduced in Ruby 3.1.0.

These changes cause issues to crop up with the Rubocop
Style/MethodDefParentheses cop which is why it's being disabled on
certain methods. This issue has been logged with the Rubocop team and
hopefully will be fixed soon.